first commit -small program

+++ small program/mistakeTest.py
@@ -1,13 +1,13 @@
 
 
 try:    
-    width: float=input("Enter rectangle width: ")
-    length :float=input("Enter rectangle length: ")
+    width : int =input("Enter rectangle width: ")
+    length : int =input("Enter rectangle length: ")
     
     if width == length:
         exit("That look like a square.")
 
-    area :int = width * length
+    area : int = width*length
     print(area)
 except ValueError:
     print('Enter a number,please.')
